% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_lst.R
\name{aes_lst}
\alias{aes_lst}
\title{aes_lst}
\usage{
aes_lst(
  est = TRUE,
  conf = TRUE,
  line = FALSE,
  mp = TRUE,
  obs = FALSE,
  shape = c(19, NA, NA, 17, NA, NA),
  size = c(1, NA, 0.2, 1, NA, NA),
  col = c("dodgerblue", NA, "grey50", "orange", NA, NA),
  fill = c("dodgerblue", "dodgerblue", NA, "orange", "grey60", "white"),
  alpha = c(1, 0.4, 1, 1, 1, NA),
  mp_pal = hcl.colors(100, palette = "Cividis", rev = FALSE),
  id_pal = "Harmonic",
  date_pal = hcl.colors(100, palette = "Viridis", rev = FALSE)
)
}
\arguments{
\item{est}{logical; turns on estimated locations (default = TRUE)}

\item{conf}{logical; turns on estimated confidence ellipses. Default varies
depending on whether behavioural index is being mapped &/or if single versus
multiple tracks are being mapped.}

\item{line}{logical; turns on estimated track line(s) (default varies)}

\item{mp}{logical; turns on move persistence index (default = TRUE, if present in model fit)}

\item{obs}{logical; turns on observed locations (default = FALSE)}

\item{shape}{gpplot2 shape value (integer: 0, 25) for estimated & observed
locations}

\item{size}{ggplot2 size value for estimated locations & track lines, and
observed locations}

\item{col}{colour for estimated locations and track lines, and observed
locations}

\item{fill}{fill colour for estimated locations & confidence ellipses,
observed locations, land polygons, and water}

\item{alpha}{transparency for specified fills/colours}

\item{mp_pal}{continuous colour palette for move persistence values}

\item{id_pal}{discrete colour palette for track id's}

\item{date_pal}{continuous colour palette for displaying date along track}
}
\value{
a named list, with 9 elements, of map components and aesthetics
\enumerate{
\item elements 1-5 - map components: \code{est}, \code{conf}, \code{line}, \code{mp}, \code{obs}
\item element 6: a data.frame named \code{df} containing ggplot2 aesthetics: \code{shape},
\code{size}, \code{col}, \code{fill}, and \code{alpha}. \code{df} has 6 rows, corresponding to the map
features:  estimated locations, confidence ellipses, track lines, observed
locations, land polygons, and water
\item element 7: \code{mp_pal}
\item element 8: \code{id_pal}
\item element 9: \code{date_pal}
}
}
\description{
set aesthetics as a named list for: 1) mapping i) estimated
locations; ii) estimated confidence ellipses; iii) estimated track lines;
iv) observed locations; v) land masses; vi) water bodies; and 2) colour
palettes for i) behavioural indices or ii) individual animal tracks
}
\examples{
# generate custom aes list
aes <- aes_lst(conf = FALSE, mp_pal = hcl.colors(n=100, palette = "RdBu"))

# modify aesthetics
aes$df$size[1] <- 1.5
aes$df$fill[6] <- grey(0.9)

}
