% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniMotum-package.R
\name{aniMotum-package}
\alias{aniMotum-package}
\alias{aniMotum}
\title{\pkg{aniMotum}}
\description{
fit Continuous-Time Random Walk and Correlated Random Walk state-space models to filter Argos Least Squares or Kalman Filter location data
}
\references{
Jonsen ID, Grecian WJ, Phillips L, et al. (2023) \code{aniMotum}, an R package for animal movement data: rapid quality control, behavioural estimation and simulation. Methods in Ecology and Evolution 14:806-816.

Jonsen ID, Patterson TA, Costa DP, et al. (2020) A continuous-time state-space model for rapid quality-control of Argos locations from animal-borne tags. Movement Ecology 8:31.

Jonsen ID, McMahon CR, Patterson TA, et al. (2019) Movement responses to environment: fast inference of variation among southern elephant seals with a mixed effects model. Ecology. 100(1):e02566.
}
\seealso{
Useful Links:
\itemize{
\item \url{https://ianjonsen.github.io/aniMotum/}
\item \url{https://github.com/ianjonsen/aniMotum/}
\item Report bugs/issues at \url{https://github.com/ianjonsen/aniMotum/issues/}
}
}
\author{
\strong{Maintainer:} Ian Jonsen \href{mailto:ian.jonsen@mq.edu.au}{ian.jonsen@mq.edu.au} \href{https://orcid.org/0000-0001-5423-6076}{ORCID}

\strong{Contributor:} Toby Patterson \href{https://orcid.org/0000-0002-7150-9205}{ORCID}

\strong{Contributor:} W. James Grecian \href{https://orcid.org/0000-0002-6428-719X}{ORCID}
}
\keyword{aniMotum}
