% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{argos_lc}
\alias{argos_lc}
\title{simulate Argos error classes for LS data}
\usage{
argos_lc(N)
}
\arguments{
\item{N}{number of location class values to be randomly sampled}
}
\description{
uses data from: 1) all IMOS reprocessed KF data (2012 - 2015);
2) LESE, SESE, HBTU KF data in Jonsen et al. 2020 Move Ecol to specify
Argos location quality class proportions
}
\keyword{internal}
