% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{ellie}
\alias{ellie}
\title{Southern elephant seal Argos satellite data (1 individual, sub-sampled for testing speed)}
\format{
.RData
}
\description{
Example elephant seal Argos tracking data. Data were sourced from
the Integrated Marine Observing System (IMOS) Sourced
from the Australian Integrated Marine Observing System (IMOS) deployments at
Davis Research Station, Antarctica and are publicly available (http:// imos.aodn.org.au).
IMOS is supported by the Australian Government through the National Collaborative
Research Infrastructure Strategy and the Super Science Initiative.
}
\keyword{data}
