% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{ellp.par}
\alias{ellp.par}
\title{simulate Argos ellipse params & sample errors from multivariate normal}
\usage{
ellp.par(lc)
}
\arguments{
\item{lc}{Argos location classes}
}
\description{
uses internal extdata/error_ellipse.RDS params ordered by lc
to sample smaj, smin from log-normals & eor from a von Mises, then convert
ellipse params to covariance matrix and sample x,y errors from multivariate
normal. Returns both ellipse params and x,y errors
}
\keyword{internal}
