% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_ssm.R
\name{elps}
\alias{elps}
\title{generate error ellipses from x,y coordinates, semi-major, semi-minor axes and ellipse orientation}
\usage{
elps(x, y, a, b, theta = 90, conf = TRUE)
}
\arguments{
\item{x}{x-coordinate, usually in projected units (km)}

\item{y}{y-coordinate, usually in projected units (km)}

\item{a}{ellipse semi-major axis (km)}

\item{b}{ellipse semi-minor axis (km)}

\item{theta}{ellipse orientation from north (degrees)}
}
\description{
generate error ellipses from x,y coordinates, semi-major, semi-minor axes and ellipse orientation
}
\keyword{internal}
