% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{epar}
\alias{epar}
\title{simulate Argos ellipse params & sample errors from multivariate normal}
\usage{
epar(lc)
}
\arguments{
\item{lc}{Argos location classes}
}
\description{
uses internal extdata/error_ellipse.RDS params ordered by lc 
to sample smaj, smin from log-normals & eor froma von Mises, then convert
ellipse params to covariance matrix and sample x,y errors from multivariate
normal. Returns both ellipse params and x,y errors
}
