% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_newton.R
\name{extra_newton}
\alias{extra_newton}
\title{extra_newton}
\usage{
extra_newton(x, n = 1)
}
\arguments{
\item{x}{a \code{ssm} fit object with class \code{ssm_df}}

\item{n}{number of extra Newton optimization steps to take}
}
\value{
an updated \code{ssm} fit object
}
\description{
Run extra optimization on an already fitted aniMotum object
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# run 3 extra Newton steps
test_extra <- extra_newton(x = fit, n = 3)

}
