% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{fit}
\alias{fit}
\title{foieGras example fit object}
\format{.RData}
\description{
Example foieGras fit object, using ellie example data and the
following call: fit <- fit_ssm(ellie, model="rw", time.step=24). This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.

Example osar object, using rope example data and the
following call: fit_res <- osar(fit). This example
osar object is included purely to speed up examples where an osar object is required
but estimating residuals is not the focus of the example.
}
\keyword{data}
