% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{fit_res}
\alias{fit_res}
\title{foieGras example osar (One-Step-Ahead residuals) object}
\format{.RData}
\description{
Example osar object, using rope example data and the
following call: fit_res <- osar(fit). This example
osar object is included purely to speed up examples where an osar object is required
but estimating residuals is not the focus of the example.
}
\keyword{data}
