% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{fite}
\alias{fite}
\title{foieGras example fit object}
\format{.RData}
\description{
Example foieGras fit object, using ellie example data and the
following call: fite <- fit_ssm(ellie, vmax=4, model="crw", time.step=24). This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
