% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{fitr}
\alias{fitr}
\title{foieGras example fit object}
\format{.RData}
\description{
Example foieGras ssm fit object, using rope example data and the
following call: fitr <- fit_ssm(rope, vmax=10, model="crw", time.step=12). This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
