% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{fmp}
\alias{fmp}
\title{foieGras example mpm fit object}
\format{.RData}
\description{
Example foieGras fit object, using rope example data and the
following call: fmp <- fitr %>% grab(., "p", as_sf = FALSE) %>% 
select(id,date,lon,lat) %>% 
fit_mpm(., model="mpm") 
This example fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
