% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmpm.R, R/foieGras-package.R
\docType{data}
\name{fmpm}
\alias{fmpm}
\title{fit the move persistence model to regularised location data}
\format{.RData}
\usage{
fmpm(x, model = c("mpm", "jmpm"), optim = c("nlminb", "optim"),
  verbose = FALSE, control = NULL, inner.control = NULL)
}
\arguments{
\item{x}{temporally regularised location data, eg. output from \code{fit_ssm}}

\item{model}{specify whether MPM is to be fit with unpooled ("mpm") or pooled ("jmpm") RW variance(s).}

\item{optim}{numerical optimizer to be used ("nlminb" or "optim")}

\item{verbose}{report progress during minimization}

\item{control}{list of control parameters for the outer optimization (type ?nlminb or ?optim for details)}

\item{inner.control}{list of control settings for the inner optimization
(see ?TMB::MakeADFUN for additional details)}
}
\description{
generates initial values for model parameters and unobserved gamma's;
structures data and initial values for C++ \code{TMB} template;
fits move persistence model; minimises the joint log-likelihood via the selected
optimizer (\code{nlminb} or \code{optim}); structures and passes output
object to \code{fit_mpm}

Example foieGras fit object, using ssm_fits example data and the
following call: fmp <- fssm %>% grab(., "p", as_sf = FALSE) %>% 
select(id,date,lon,lat) %>% 
fit_mpm(., model="jmpm") 
This example fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\details{
called by \code{fit_mpm}, see ?fit_mpm.
}
\keyword{data}
