% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-deprecated.R
\name{pluck}
\alias{pluck}
\title{Deprecated functions.}
\usage{
pluck(...)
}
\arguments{
\item{...}{ignored}
}
\description{
\code{pluck} has been deprecated to avoid conflict with purrr::pluck. Instead use
\code{grab}.
}
