% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{package}
\name{foieGras-package}
\alias{foieGras-package}
\alias{foieGras}
\title{\pkg{foieGras}}
\description{
fit Continuous-Time Random Walk and Correlated Random Walk state-space models to filter Argos Least Squares or Kalman Filter location data
}
\author{
Ian Jonsen, Toby Patterson

 @seealso fit_ssm
 @references Jonsen I, McMahon CR, Patterson TA, Auger-Methe M, Harcourt R, Hindell MA, Bestley S (2019) Movement responses to environment: fast inference of variation among southern elephant seals with a mixed effects model. Ecology 100:e02566

 @keywords foieGras
 @importFrom utils data
 @importFrom dplyr group_by do rowwise %>% ungroup select mutate slice
 @importFrom dplyr distinct arrange filter left_join lag full_join bind_cols
 @importFrom dplyr summarise
 @importFrom tibble as_tibble
 @importFrom lubridate ymd_hms
 @importFrom sf st_as_sf st_set_crs st_transform st_is_longlat st_crs
 @importFrom sf st_coordinates st_geometry<- st_bbox st_cast
 @importFrom argosfilter sdafilter
 @importFrom TMB MakeADFun sdreport newtonOption
 @importFrom stats approx cov sd predict nlminb optim na.omit
 @importFrom ggplot2 ggplot geom_point geom_path aes ggtitle theme_bw
 @importFrom ggplot2 theme element_blank geom_sf xlim ylim unit
 @importFrom ggplot2 element_text scale_colour_viridis_c
 @importFrom gridExtra grid.arrange
 @importFrom grDevices extendrange grey
}
