% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Coerce input data into expected \code{foieGras} format}
\usage{
format_data(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("lonerr", "laterr"),
  tz = "UTC"
)
}
\arguments{
\item{x}{input data}

\item{id}{the name (as a quoted character string) of id variable: a unique
identifier for individual (animal) track data sets.}

\item{date}{the name (as a quoted character string)of the date/time variable:
date and time (as YYYY-MM-DD HH:MM:SS) of each observation.}

\item{lc}{the name (as a quoted character string) of the location quality class
variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z").
Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS)
and other data types.}

\item{coord}{the names (as quoted character strings) of the location coordinate
variables: defaults are c("lon","lat"), but could also be c("x","y") for planar
coordinates; or if input data is an \code{sf} object then "geometry". If input
data is an \code{sf} object then \code{coord} is set to "geometry" by default.}

\item{epar}{the names (as quoted character strings) of the Argos error ellipse
parameters: defaults are "smaj" (ellipse semi-major axis),
"smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if
these variables are missing from the input data.}

\item{sderr}{the names (as quoted character strings) of provided standard
errors in longitude and latitude: defaults are "lonerr", "laterr". Typically,
these are only provided for processed light-level geolocation data. Ignored if
these variables are missing from the input data.}

\item{tz}{the timezone the applies to the data/time variable if they are not
in \code{tz = 'UTC'}. A list of valid timezone names can be viewed via \code{OlsonNames()}}
}
\value{
a data.frame or sf-tibble of input data in expected foieGras format.
Additional columns required by \code{fit_ssm()}, if missing, will be added to the
formatted tibble: \code{smaj}, \code{smin}, \code{eor}, \code{lonerr}, and \code{laterr}.
}
\description{
format data by mapping supplied variable names to those expected by
\code{fit_ssm()}, and ensuring variables are put into the expected order. Can be
run manually by user as a data pre-processing step prior to calling \code{fit_ssm()}
or can be called automatically by \code{fit_ssm()}. In the latter case, any custom
variable names must be declared as arguments to \code{fit_ssm()}; see examples, below.
}
\examples{
## as a data pre-processing step
data(sese2_n)
d <- format_data(sese2_n, date = "time", coord = c("longitude","latitude"), 
tz = "America/Halifax")
fit <- fit_ssm(d, model = "crw", time.step = 24)

## called automatically within fit_ssm()
fit <- fit_ssm(sese2_n, date = "time", coord = c("longitude", "latitude"), 
tz = "America/Halifax", model = "crw", time.step = 24)
}
