% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{grab}
\alias{grab}
\title{grab tibble's by name from a foieGras model object}
\usage{
grab(x, what = "fitted", as_sf = FALSE, normalise = FALSE, group = FALSE)
}
\arguments{
\item{x}{a \code{foieGras} ssm or mpm model object}

\item{what}{the tibble to be grabbed; either \code{fitted}, \code{predicted},
\code{rerouted} (ssm only), or \code{data} (single letters can be used).}

\item{as_sf}{logical; if FALSE (default) then return a tibble with
un-projected lonlat coordinates, otherwise return an sf tibble. Ignored if x
is an mpm model object.}

\item{normalise}{logical; if output includes a move persistence estimate,
should \code{g} (the move persistence index) be normalised to have minimum = 0 and
maximum = 1 (default = FALSE). Note, this normalisation is not applied to the
standard errors of the logit-scale move persistence estimates (\code{logit_g},
\code{logit_g.se}).}

\item{group}{logical; should \code{g} be normalised among individuals as a group,
a 'relative g', or to individuals separately to highlight regions of lowest
and highest move persistence along single tracks (default = FALSE).}
}
\value{
a tibble with all individual tibble's appended
}
\description{
\code{grab()} lets you obtain \code{fitted}, \code{predicted}, \code{rerouted} or
\code{data} tibble's from a compound tibble created when fitting to multiple
individual data sets. The specified tibble's are appended to a single output
tibble.
}
\examples{
## generate an ssm fit object
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
control = ssm_control(verbose = 0))

## grab predicted values as an un-projected tibble
preds <- grab(xs, what = "predicted")

}
