% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{join an mpm-estimated behavioural index to ssm-predicted locations}
\usage{
join(
  ssm,
  mpm,
  what.ssm = "predicted",
  as_sf = FALSE,
  normalise = FALSE,
  group = FALSE
)
}
\arguments{
\item{ssm}{a \code{foieGras} ssm fitted model object}

\item{mpm}{a \code{foieGras} mpm fitted model object}

\item{what.ssm}{specifies whether ssm \code{predicted} or \code{fitted} values are to be extracted}

\item{as_sf}{logical; if FALSE then return a tibble with un-projected lonlat
coordinates, otherwise return an sf tibble}

\item{normalise}{logical; if output includes a move persistence estimate,
should g (the move persistence index) be normalised to have minimum = 0 and
maximum = 1 (default = FALSE).}

\item{group}{logical; should g be normalised among individuals as a group,
a 'relative g', or separately to highlight regions of lowest and highest move
persistence along a track (default = FALSE).}
}
\value{
a single tbl with all individuals
}
\description{
\code{join()} joins ssm-predicted locations and mpm-estimated behavioural index into a single tibble. If the ssm-predicted tibble is a projected sf object then the output of join will also be an sf object (default). This can be avoided by using \code{as_sf = FALSE}.
}
\examples{
## load example foieGras fit objects (to save time)
## generate a ssm fit object
xs <- fit_ssm(ellie, spdf=FALSE, model = "rw", time.step=24, control = ssm_control(verbose = 0))
xm <- fit_mpm(xs, what = "p", model = "mpm")

## join predicted values as an un-projected tibble
xsm <- join(xs, xm)
xsm
}
