% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.r
\name{map}
\alias{map}
\title{map}
\usage{
map(x, est = c("fitted", "predicted"), se = TRUE, outlier = FALSE)
}
\arguments{
\item{x}{a foieGras fitted object}

\item{est}{specify which location estimates to map: fitted or predicted}

\item{se}{include 2 * SE on time-series plots (SE's currently not displayed when proj = "ll")}

\item{outlier}{include outlier observations identified during prefilter-ing}
}
\description{
map foieGras fitted or predicted locations
}
