% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_multi_track_base.R
\name{map_multi_track_base}
\alias{map_multi_track_base}
\title{map_multi_track_base}
\usage{
map_multi_track_base(
  map_type,
  obs_sf,
  conf_sf,
  line_sf,
  loc_sf,
  by.id,
  by.date,
  extents,
  buffer,
  aes,
  silent,
  ...
)
}
\arguments{
\item{map_type}{coastline data to be used in maps}

\item{obs_sf}{observed location geometry}

\item{conf_sf}{confidence ellipses around estimated locations, if specified}

\item{line_sf}{track line, if specified}

\item{loc_sf}{estimated location geometry}

\item{by.id}{colour estimated locations by id (logical)}

\item{by.date}{colour estimated locations by date (logical)}

\item{extents}{map extents}

\item{buffer}{distance (in km) to buffer locations for subsetting land
polygons (default = 10000). If map extents are expanded by many factors then
the buffer distance may need to be increased, otherwise this should not be
used.}

\item{aes}{a list of map aesthetics (size, shape, col, fill, alpha) to
be applied, in order, to: 1) estimated locations,; 2) confidence ellipses;
3) track lines; 4) observed locations; 5) land regions; 6) water regions}

\item{silent}{logical; map silently (default = FALSE)}

\item{...}{additional arguments passed to \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile}}
}
\description{
map foieGras-estimated locations from multiple tracks without
a behavioural index
}
\keyword{internal}
