% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_single_track_base.R
\name{map_single_track_base}
\alias{map_single_track_base}
\title{map_single_track_base}
\usage{
map_single_track_base(
  map_type,
  obs_sf,
  conf_sf,
  line_sf,
  loc_sf,
  by.date,
  extents,
  aes,
  ...
)
}
\arguments{
\item{map_type}{coastline data to be used in maps}

\item{obs_sf}{observed location geometry}

\item{conf_sf}{confidence ellipses around estimated locations, if specified}

\item{line_sf}{track line, if specified}

\item{loc_sf}{estimated location geometry}

\item{by.date}{colour estimated locations by date (logical)}

\item{extents}{map extents}

\item{aes}{a tibble of map aesthetics (size, shape, col, fill, alpha) to
be applied, in order, to: 1) estimated locations,; 2) confidence ellipses;
3) track lines; 4) observed locations; 5) land regions; 6) water regions}

\item{...}{additional arguments passed to \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile}}
}
\description{
map foieGras-estimated locations from a single track without
a behavioural index
}
\keyword{internal}
