% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osar.r
\name{osar}
\alias{osar}
\title{calculate one-step-ahead (prediction) residuals from a \code{foieGras} fit}
\usage{
osar(x, method = "oneStepGaussianOffMode", ...)
}
\arguments{
\item{x}{a compound \code{fG} tbl fit object}

\item{method}{method to calculate prediction residuals (default is "oneStepGaussianOffMode"; see `?TMB::oneStepPrediction` for details)}
}
\description{
calculate one-step-ahead (prediction) residuals from a \code{foieGras} fit
}
\details{
One-step-ahead residuals are useful for assessing goodness-of-fit in latent variable models. This is a wrapper function for TMB::oneStepPredict (beta version)
}
\examples{
## see summary fit output
## load example foieGras fit object (to save time)
data(fit)
osar(fit)

}
\references{
Thygesen, U. H., C. M. Albertsen, C. W. Berg, K. Kristensen, and A. Neilsen. 2017. Validation of ecological state space models using the Laplace approximation. Environmental and Ecological Statistics 24:317–339.
}
