% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_add_emf.R
\name{pf_add_emf}
\alias{pf_add_emf}
\title{Add error multiplication factors for Argos LS data}
\usage{
pf_add_emf(x, emf)
}
\arguments{
\item{x}{data from \code{pf_sf_project}}

\item{emf}{optionally supplied data.frame of error multiplication factors
for Argos location quality classes. Relevant to Argos LS and GPS data only}
}
\description{
Adds location error multiplication factors based on Argos location
class (for type LS) & finalises prefiltered data for use by \code{sfilter()}
}
\keyword{internal}
