% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_dup_dates.R
\name{pf_dup_dates}
\alias{pf_dup_dates}
\title{Find observations with duplicate dates}
\usage{
pf_dup_dates(x, min.dt)
}
\arguments{
\item{x}{input data from \code{format_data()}}

\item{min.dt}{minimum allowable time difference in s between observations;
\code{dt < min.dt} will be ignored by the SSM}
}
\description{
Find observations with duplicate dates
}
\keyword{internal}
