% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_obs_type.R
\name{pf_obs_type}
\alias{pf_obs_type}
\title{Determine observation type: LS, KF, GPS, or GLS}
\usage{
pf_obs_type(x)
}
\arguments{
\item{x}{data from \code{pf_dup_dates()}}
}
\description{
Determine observation type: LS, KF, GPS, or GLS
}
\keyword{internal}
