% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_sda_filter.R
\name{pf_sda_filter}
\alias{pf_sda_filter}
\title{Use \code{trip::sda} to identify extreme locations}
\usage{
pf_sda_filter(x, spdf, vmax, ang, distlim)
}
\arguments{
\item{x}{data from \code{pf_obs_type()}}

\item{spdf}{turn speed filter on/off (logical; default is TRUE)}

\item{vmax}{max travel rate (m/s)}

\item{ang}{angles of outlier location "spikes" (default is \code{c(15,25)} deg);
\code{ang = NA} turns off \code{trip::sda} filter in favour of
\code{trip::speedfilter}}

\item{distlim}{lengths of outlier location "spikes" in km (default is
\code{c(2.5, 5)} m); \code{distlim = NA} turns off \code{trip::sda} filter
in favour of \code{trip::speedfilter}. Either \code{ang = NA} or
\code{distlim = NA} are sufficient.}
}
\description{
Use \code{trip::sda} to identify extreme locations
}
\details{
\code{trip::sda} is a vectorized implementation of the Argos filter
presented in Freitas et al. (2008) Mar Mamm Sci 24:315-325. This function
checks for errors returned by \code{trip::sda} and falls back to using the simpler
\code{trip::speedfilter} if an error is returned.
}
\keyword{internal}
