% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG.R
\name{plot.fG}
\alias{plot.fG}
\title{plot}
\usage{
\method{plot}{fG}(x, what = c("fitted", "predicted"), outlier = FALSE,
  ...)
}
\arguments{
\item{x}{a fG compound tibble}

\item{what}{specify which location estimates to display on time-series plots: fitted or predicted}

\item{outlier}{include all extreme outliers flagged by prefilter in plots (logical)}

\item{...}{additional arguments to be ignored}
}
\description{
visualise multiple fits from an fG compound tibble
}
\examples{
## load example foieGras fit object (to save time)
data(fit)
plot(fit$ssm[[1]])

}
