% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_osar.R
\name{plot.fG_osar}
\alias{plot.fG_osar}
\title{plot}
\usage{
\method{plot}{fG_osar}(x, type = c("qqnorm", "histogram", "boxplot"),
  bw = 0.5, ...)
}
\arguments{
\item{x}{a \code{foieGras osar} object with class `fG_osar`}

\item{type}{type of residual plot to generate; either qqnorm (default), histogram or boxplot}

\item{bw}{binwidth for histogram plots (see ggplot2::geom_histogram for details)}

\item{...}{additional arguments to be ignored}
}
\description{
plot One-Step-Ahead (prediction) residuals from a \code{foieGras osar} object
}
\examples{
## load example osar output (to save time)
data(fite_res)
plot(fite_res, "qq")
plot(fite_res, "hist")

}
