% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_simfit.R
\name{plot.fG_simfit}
\alias{plot.fG_simfit}
\title{plot}
\usage{
\method{plot}{fG_simfit}(
  x,
  type = c("lines", "points", "both"),
  ext = c("hemi", "tracks"),
  ncol = 1,
  pal = "Zissou1",
  ...
)
}
\arguments{
\item{x}{a \code{foieGras} simulation data.frame with class \code{fG_simfit}}

\item{type}{plots tracks as "line", "points" or "both" (default).}

\item{ext}{map extent for plotting, either "hemi" (default) for hemisphere, or
"tracks" to zoom in on track extents. In the latter case, the projection is Mercator.}

\item{ncol}{number of columns to arrange multiple plots}

\item{pal}{\code{hcl.colors} palette to use (default: "Zissou1"; type \code{hcl.pals()} for options)}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of simulated tracks.
}
\description{
visualize tracks simulated from a foieGras model fit
}
\examples{
fit <- fit_ssm(sese1, vmax = 4, model = "crw", time.step = 72)
trs <- simfit(fit, what = "p", reps = 2)
plot(trs, type = "b")

}
