% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_ssm.R
\name{plot.fG_ssm}
\alias{plot.fG_ssm}
\title{plot}
\usage{
\method{plot}{fG_ssm}(x, what = c("fitted", "predicted"), type = 1,
  ncol = 1, ...)
}
\arguments{
\item{x}{a \code{foieGras} ssm fit object with class `fG_ssm`}

\item{what}{specify which location estimates to display on time-series plots: fitted or predicted}

\item{type}{of plot to generate: 1-d time series for lon and lat separately (type = 1, default) or 2-d track plot (type = 2)}

\item{ncol}{number of columns to use for faceting. Default is ncol = 1 but this may be increased for multi-individual fit objects}

\item{...}{additional arguments to be ignored}
}
\value{
a ggplot object with either: (type = 1) 1-D time series of fits to data, 
separated into x and y components (units = km) with prediction uncertainty ribbons (2 x SE); 
or (type = 2) 2-D fits to data (units = km)
}
\description{
visualize multiple fits from an fG compound tibble
}
\examples{
## load example foieGras fit object (to save time)
data(fssm)
plot(fssm, what = "f", type = 1)
plot(fssm, what = "p", type = 2)

}
