% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.foieGras.R
\name{plot.foieGras}
\alias{plot.foieGras}
\title{plot}
\usage{
\method{plot}{foieGras}(m, what = c("fitted", "predicted"), se = TRUE)
}
\arguments{
\item{m}{a foieGras fitted object}

\item{what}{specify which location estimates to display on time-series plots: fitted or predicted}

\item{se}{include 2 * SE on time-series plots (SE's currently not displayed when proj = "ll")}
}
\description{
Visualise foieGras SSM fits to track data
}
