% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.osar.R
\name{plot.osar}
\alias{plot.osar}
\title{plot}
\usage{
\method{plot}{osar}(x, type = c("qqnorm", "histogram", "boxplot"), ...)
}
\arguments{
\item{x}{a compound fG tibble or an individual foieGras fitted object}

\item{type}{type of residual plot to generate; either qqnorm (default), histogram or boxplot}

\item{...}{additional arguments to be ignored}
}
\description{
plot One-Step-Ahead (prediction) residuals from a foieGras fit
}
\examples{
## load example osar output (to save time)
data(fit_osar)
plot(osar, "qq")

}
