% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.R
\name{plot.sim}
\alias{plot.sim}
\title{plot}
\usage{
\method{plot}{sim}(x, error = FALSE, pal = "Cividis", rev = FALSE, col = TRUE, ...)
}
\arguments{
\item{x}{a \code{foieGras} simulation data.frame with class \code{sim}}

\item{error}{logical, plot locations with error (TRUE) or without. Ignored in 1-D time-series plots}

\item{pal}{hcl.colors palette to use (default: "Cividis"); type \code{hcl.pals()} for options}

\item{rev}{reverse direction of colour palette; logical (default = FALSE)}

\item{col}{colour data points by speed; logical (default = TRUE)}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of simulated tracks. Can be rendered all on a single page (pages = 1) or on separate pages (pages = 0).
}
\description{
visualize simulated tracks from a sim data.frame
}
\examples{
tr <- sim(N=100, model = "crw")
plot(tr, error = TRUE)

}
