% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim_fit.R
\name{plot.sim_fit}
\alias{plot.sim_fit}
\title{plot}
\usage{
\method{plot}{sim_fit}(
  x,
  type = "points",
  zoom = TRUE,
  ncol = 1,
  hires = FALSE,
  ortho = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{aniMotum} simulation data.frame with class \code{sim_fit}}

\item{type}{plots tracks as "line", "points" or "both" (default).}

\item{zoom}{logical; should map extent be defined by track extent (TRUE; default) or
should global map be drawn (FALSE).}

\item{ncol}{number of columns to arrange multiple plots}

\item{hires}{logical; use high-resolution coastline data. Attempts to use
high-res coastline data via \link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries} with \code{scale = 10},
if the \code{rnaturalearthhires} data package is installed. This extends the
plot rendering time so is set to FALSE by default, in which case
\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries} with \code{scale = 50} data are used.}

\item{...}{additional arguments to be ignored}

\item{or}{orientation of projected map, default is to centre on
start of fitted track (ignored if \code{mapproj} package is not installed). This argument
is deprecated and, if specified, will be ignored with a warning.}
}
\value{
Plots of simulated tracks.
}
\description{
visualize tracks simulated from a \code{aniMotum} model fit
}
\examples{
fit <- fit_ssm(ellie, model = "crw", time.step = 24)
trs <- sim_fit(fit, what = "p", reps = 2)
plot(trs, type = "b")

}
