% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck.R
\name{pluck}
\alias{pluck}
\title{Pluck tbl's by name from a foieGras fitted model object}
\usage{
pluck(fitobj, what)
}
\arguments{
\item{what}{the tbl to be plucked; either `fitted`, `predicted`, or `data` (single letters can be used)}

\item{fitobj}{a foieGras fitted model object}
}
\value{
a tbl with all individual tbl's appended
}
\description{
`pluck()` lets you obtain `fitted`, `predicted`, or `data` tbl's from
a compound tbl created when fitting to multiple individual data sets. The specified tbl's
are appended to a single output tbl.
}
\examples{
\dontrun{
data("ellie")
ellie.fit <- fit_ssm(ellie, model = "crw", time.step = 6)

pred <- pluck(ellie.fit, "predicted")
}
}
