% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.R
\name{prefilter}
\alias{prefilter}
\title{Prepare Argos data for fitting state-space model}
\usage{
prefilter(
  x,
  vmax = 5,
  ang = c(15, 25),
  distlim = c(2500, 5000),
  spdf = TRUE,
  min.dt = 0,
  emf = NULL,
  prj = NULL
)
}
\arguments{
\item{x}{input data, must have 5 (LS), or 8 (KF) columns (see details)}

\item{vmax}{max travel rate (m/s)}

\item{ang}{angles of outlier location "spikes" (default is \code{c(15,25)} deg);
\code{ang = NA} turns off \code{sda} filter in favour of
\code{speedfilter}}

\item{distlim}{lengths of outlier location "spikes" in km (default is
\code{c(2.5, 5)} m); \code{distlim = NA} turns off \code{sda} filter
in favour of \code{speedfilter}. Either \code{ang = NA} or
\code{distlim = NA} are sufficient.}

\item{spdf}{turn speed filter on/off (logical; default is TRUE)}

\item{min.dt}{minimum allowable time difference in s between observations;
\code{dt < min.dt} will be ignored by the SSM. Default is NA: all time
differences > 0 are allowed}

\item{emf}{optionally supplied data.frame of error multiplication factors
for Argos location quality classes. see Details}

\item{prj}{optional proj4 string, passed to \code{pf_sf_project}. Used only when
the data are unprojected; an sf object carrying a projected CRS is always
respected}
}
\value{
an sf object with all observations passed from \code{data} and the
following appended columns
\item{\code{keep}}{logical indicating whether observation should be ignored
by \code{sfilter} (FALSE)}
\item{\code{obs.type}}{flag indicating whether KF or LS measurement model
applies}
\item{\code{emf_x}}{error multiplication factors for \code{x} direction}
\item{\code{emf_y}}{error multiplication factors for \code{y} direction}
\item{\code{geometry}}{sf POINT object giving \code{x,y} coordinates in km}
}
\description{
\code{prefilter} does the following:
(1) flags (\code{keep = FALSE}) observations with duplicate
dates & removes subsequent observations occurring within min.dt s of one another
(2) determines Argos data type (LS, KF, G, or GL);
(3) uses \code{sda} filter to identify extreme locations. \code{sda} is a
fast, vectorized version of the now CRAN-archived \code{argosfilter::sdafilter}.
\code{sda} is a native implementation of that found in the currently archived \code{trip}
package (by MD Sumner: https://github.com/Trackage/trip).
(4) projects lonlat coords to mercator x,y coords (in km) & shifts longitudes
that straddle -180,180 to 0,360 and vice-versa;
(5) adds location error multiplication factors based on Argos location
class (for type LS);
}
\details{
called by \code{fit_ssm}.
}
\keyword{internal}
