% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presid.r
\name{presid}
\alias{presid}
\title{obtain one-step-ahead (prediction) residuals from a \code{foieGras} fit}
\usage{
presid(x, method = "oneStepGaussianOffMode", parallel = TRUE)
}
\arguments{
\item{x}{a compound \code{fG} tbl or a \code{foieGras} individual fit object}

\item{method}{method to calculate prediction residuals (default is "oneStepGaussianOffMode"; see `?TMB::oneStepPrediction` for details)}

\item{parallel}{run TMB::oneStepPredict in parallel (default is TRUE; requires \code{parallel} package)}
}
\description{
obtain one-step-ahead (prediction) residuals from a \code{foieGras} fit
}
\examples{
## see summary fit output
## load example foieGras fit object (to save time)
data(fit)
presid(fit$ssm[[1]])

}
