% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.foieGras.R
\name{print.foieGras}
\alias{print.foieGras}
\title{print \code{foieGras} object summary information}
\usage{
\method{print}{foieGras}(x, ...)
}
\arguments{
\item{x}{a \code{ctrw} fit object}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
print \code{foieGras} object summary information
}
\examples{
## see summary fit output
data(ellie)
fit <- fit_ssm(ellie, model="rw", time.step=24)
fit

}
