% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jssm_df.R
\name{print.jcrw_ssm}
\alias{print.jcrw_ssm}
\title{print one individual's part of a joint crw fit}
\usage{
\method{print}{jcrw_ssm}(x, ...)
}
\arguments{
\item{x}{a single individual's component of a joint \code{jcrw} fit}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
print one individual's part of a joint crw fit
}
