% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ssm.R
\name{print.ssm}
\alias{print.ssm}
\alias{print}
\title{print \code{foieGras} fit object summary information}
\usage{
\method{print}{ssm}(x, ...)
}
\arguments{
\item{x}{a foieGras ssm fit object}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
print \code{foieGras} fit object summary information
}
\examples{
## see summary fit output
## load example foieGras ssm fit object (to save time)
data(xs)
xs$ssm[[1]]

}
