% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.jssm_df.R
\name{print.summary.jssm_df}
\alias{print.summary.jssm_df}
\title{print a joint ssm fit summary}
\usage{
\method{print}{summary.jssm_df}(x, ...)
}
\arguments{
\item{x}{a \code{summary.jssm_df} object}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
print a joint ssm fit summary
}
