% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{res}
\alias{res}
\title{foieGras example osar residuals object}
\format{
.RData
}
\description{
Example foieGras osar residuals object. This example osar
residuals object is included purely to speed up vignette build.
}
\keyword{data}
