% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{sese2}
\alias{sese2}
\title{Southern elephant seal Argos satellite data (2 highly sub-sampled individuals)}
\format{
.RData
}
\description{
Example elephant seal Argos tracking data, highly sub-sampled.
These example data are included purely to speed up examples where a fit
object is required. Generating a fit object is preferred as storing an
example fit risks GDAL errors on platforms with older GDAL libraries
}
\keyword{data}
