% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{simulate animal tracks}
\usage{
sim(
  N = 100,
  start = list(c(0, 0), as.POSIXct(format(Sys.time(), tz = "UTC", usetz = TRUE))),
  model = c("rw", "crw", "mp"),
  vmax = 4,
  sigma = c(4, 4),
  rho_p = 0,
  D = 0.05,
  sigma_g = 1.25,
  error = c("ls", "kf"),
  tau = c(1.5, 0.75),
  rho_o = 0,
  tdist = c("reg", "gamma"),
  ts = 6,
  tpar = 1.2,
  alpha = c(0.9, 0.8)
)
}
\arguments{
\item{N}{number of time steps to simulate}

\item{start}{coordinates and datetime of start location for simulated track}

\item{model}{simulate from the \code{rw}, \code{crw} or \code{mp} process models}

\item{vmax}{maximum travel rate (m/s) of simulated animal}

\item{sigma}{a vector of process error sd's for the \code{rw} model
(ignored if \code{model != "rw"})}

\item{rho_p}{correlation parameter for \code{rw} model process covariance matrix
(ignored if \code{model != "rw"})}

\item{D}{diffusion coefficient for \code{crw} model process covariance matrix
(ignored if \code{model != "crw"})}

\item{sigma_g}{random walk sd for time-varying move persistence parameter
(ignored if \code{model != "mp"})}

\item{error}{indicates whether measurement error should mimic Argos
Least-Squares (\code{ls}) or Argos Kalman Filter (\code{kf})}

\item{tau}{vector of LS measurement error sd's (ignored if \code{error = "kf"})}

\item{rho_o}{correlation parameter for LS covariance matrix
(ignored if \code{error = "kf"})}

\item{tdist}{distribution for simulating location times (\code{reg} generates locations
at regular ts intervals, in h; \code{gamma} uses a gamma distribution to generate random
time intervals)}

\item{ts}{time interval in h}

\item{tpar}{rate parameter for the gamma distributed times, shape is take to be
\code{ts * tpar} for a mean interval of approximately \code{ts} h
(ignored if \code{tdist = "reg"})}

\item{alpha}{transition probabilities switching model versions of
\code{rw} or \code{crw} models. Probabilities are the transition matrix diagonals
(ignored if sigma has length 2 or D has length 1)}
}
\value{
a tibble is returned with columns that can include some or all of the
following, depending on the arguments used
\itemize{
\item \code{date} time as POSIXct, tz = UTC (default)
\item \code{lc} Argos location class
\item \code{lon} longitude with error
\item \code{lat} latitude with error
\item \code{x} x in km from arbitrary origin without error
\item \code{y} y in km from arbitrary origin without error
\item \code{x.err} a random deviate drawn from Argos LS or KF error distribution
\item \code{y.err} a random deviate drawn from Argos LS or KF error distribution
\item \code{smaj} Argos error ellipse semi-major axis in m (if \code{error = "kf"})
\item \code{smin} Argos error ellipse semi-minor axis in m (if \code{error = "kf"})
\item \code{eor} Argos error ellipse orientation in degrees (if \code{error = "kf"})
\item \code{u} velocity in x direction (if \code{model = "crw"}), unit = km/h
\item \code{v} velocity in y direction (if \code{model = "crw"}), unit = km/h
\item \code{b} behavioural state (if \code{model = "rw"} or \code{model = "crw"} and multiple process variances given, see examples)
\item \code{g} movement persistence - the autocorrelation between successive movements on the interval 0,1 (if \code{model = "mp"})
}
}
\description{
simulate from the \code{rw}, \code{crw}, or \code{mp} process models
to generate a set of \verb{x,y} (or \verb{lon,lat}) coordinates with or without error
from supplied input parameters.
}
\examples{
tr <- sim(N = 200, model = "crw", D = 0.1, error = "kf", tdist = "reg", ts=12)
plot(tr, error = TRUE)

tr <- sim(N = 200, model = "mp", sigma_g = 1.2, error = "ls", tau = c(2, 1.5), ts=12,
tdist = "gamma", tpar = 1.5)
plot(tr, error = TRUE, pal = "Cividis")

}
