% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_filter.R
\name{sim_filter}
\alias{sim_filter}
\title{Filter \code{simfit} simulations}
\usage{
sim_filter(trs, keep = 0.25, flag = 2)
}
\arguments{
\item{trs}{a \code{simfit} object}

\item{keep}{the quantile of flag values to retain}

\item{flag}{the similarity flag method (see \code{?simlarity_flag} for details)}
}
\value{
a \code{simfit} object containing the filtered paths
}
\description{
This function calculates the similarity between the simulations
generated by \code{simfit} and the observed path used in the \code{fG_ssm} fit,
and returns a \code{simfit} object containing the most similar tracks based on
a user specified quantile.
}
\examples{
## fit crw model to Argos LS data
fit <- fit_ssm(sese1, vmax = 4, model = "crw", time.step = 24)

## generate 10 simulated paths from ssm fit
trs <- simfit(fit, what = "predicted", reps = 10)

## filter simulations and keep paths in top 25\% of flag values
trs_f <- sim_filter(trs, keep = .25, flag = 2)
plot(trs_f)
}
\seealso{
\code{\link{similarity_flag}}
}
