% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_post.R
\name{sim_post}
\alias{sim_post}
\title{simulate from the posterior of a \code{ssm} fit, conditional on parameters and data}
\usage{
sim_post(x, reps = 1, what = "predicted")
}
\arguments{
\item{x}{a \code{ssm} fit object with class \code{ssm_df}}

\item{reps}{number of replicate tracks to simulate from an \code{ssm} model
fit object}
}
\value{
a \code{fG_sim_post} object containing the paths simulated from a
\code{ssm} fit object
}
\description{
...
}
\examples{
fit <- fit_ssm(ellie, model = "crw", time.step = 24)
trs <- sim_post(fit, reps = 3)
plot(trs)

}
