% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_var.R
\name{similarity_var}
\alias{similarity_var}
\title{Add similarity var value to a \code{sim_fit} object}
\usage{
similarity_var(track, sim_track, var, FUN = "mean", ...)
}
\arguments{
\item{track}{a dataframe containing longitude and latitudes of the observed
path used in the SSM fit}

\item{sim_track}{a dataframe containing the longitude and latitude of a single
simulated path from sim_fit}

\item{FUN}{the function used to summarise the user-supplied variable}
}
\value{
a single value representing the similarity between the two paths
}
\description{
Calculate the similarity between a single path generated by
sim_fit and the observed path used in the SSM fit. In this context,
similarity is calculated using the mean, median, variance, standard
deviation, sum, min, or max of a user-supplied variable appended to the
observed and simulated tracks.
}
\keyword{internal}
