% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{speedfilter}
\alias{speedfilter}
\title{speedfilter: filter track for speed.}
\usage{
speedfilter(x, max.speed = NULL, test = FALSE)
}
\arguments{
\item{x}{fG_format object passed from \code{sda}}

\item{max.speed}{maximum speed, in km/h}
}
\value{
logical vector, with \code{FALSE} values where the tests failed
}
\description{
Create a filter index of a track for "bad" points based only on
speed. Called from \code{sda}
}
\details{
This is an implementation based on that in the
package trip by MD Sumner (https://github.com/Trackage/trip).
}
\keyword{internal}
