% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{ssm_fits}
\alias{ssm_fits}
\title{foieGras example fit object}
\format{.RData}
\description{
Example foieGras fit object, using ellies example data and the
following call: ssm_fits <- fit_ssm(ellies, vmax=4, model="crw", time.step=48). This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
