% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ssm.R
\name{summary.ssm_df}
\alias{summary.ssm_df}
\title{object summaries}
\usage{
\method{summary}{ssm_df}(object, ...)
}
\arguments{
\item{object}{an \code{ssm_df} fit object}

\item{...}{additional arguments to be ignored}
}
\description{
return a summary of an \code{ssm_df} fit object
}
