% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{testfit}
\alias{testfit}
\title{a test foieGras fitted model object}
\format{.RData}
\description{
a foieGras fitted model object, using the ellie example data,
for package testing purposes
}
\keyword{data}
