% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{wese_sb}
\alias{wese_sb}
\title{Weddell seal Argos satellite data (1 individual)}
\format{
.RData
}
\description{
Example Weddell seal Argos tracking data, deployed at Scott Base,
Ross Island, Antarctica. This example data set is included for demonstration
purposes. Sourced from the Australian Integrated Marine Observing System (IMOS)
& the New Zealand National Institute of Water and Atmospheric Research (NIWA)
deployments at Scott Base, Antarctica and are publicly available (http:// imos.aodn.org.au).
IMOS is supported by the Australian Government through the National Collaborative
Research Infrastructure Strategy and the Super Science Initiative.
}
\keyword{data}
