% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_crw.r, R/util_rw.r
\name{wrap_lon}
\alias{wrap_lon}
\title{Utility functions}
\usage{
wrap_lon(lon, lon_min = -180)

wrap_lon(lon, lon_min = -180)
}
\description{
Utility functions

Utility functions
}
\details{
Internal function, typically not called by user

Internal function, typically not called by user
}
