% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{data}
\name{xs}
\alias{xs}
\title{foieGras example fit object}
\format{
.RData
}
\description{
Example foieGras fit object. This example
fit is included purely to speed up examples where a fit object is required
but fitting to data is not the focus of the example.
}
\keyword{data}
