% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{aes_lst}
\alias{aes_lst}
\title{aes_lst}
\usage{
aes_lst(
  est = TRUE,
  conf = TRUE,
  line = FALSE,
  mp = TRUE,
  obs = FALSE,
  shape = c(19, NA, NA, 17, NA, NA),
  size = c(1, NA, 0.2, 1, NA, NA),
  col = c("dodgerblue", NA, "grey50", "orange", NA, NA),
  fill = c("dodgerblue", "dodgerblue", NA, "orange", "grey60", "white"),
  alpha = c(1, 0.4, 1, 1, 1, NA),
  mp_pal = hcl.colors(100, palette = "Cividis", rev = FALSE),
  id_pal = "Harmonic",
  date_pal = hcl.colors(100, palette = "Viridis", rev = FALSE)
)
}
\arguments{
\item{est}{logical; turns on estimated locations (default = TRUE)}

\item{conf}{logical; turns on estimated confidence ellipses. Default varies
depending on whether behavioural index is being mapped &/or if single versus
multiple tracks are being mapped.}

\item{line}{logical; turns on estimated track line(s) (default varies)}

\item{mp}{logical; turns on move persistence index (default = TRUE, if present in model fit)}

\item{obs}{logical; turns on observed locations (default = FALSE)}

\item{shape}{gpplot2 shape value (integer: 0, 25) for estimated & observed
locations}

\item{size}{ggplot2 size value for estimated locations & track lines, and
observed locations}

\item{col}{colour for estimated locations and track lines, and observed
locations}

\item{fill}{fill colour for estimated locations & confidence ellipses,
observed locations, land masses, and water bodies}

\item{alpha}{transparency for specified fills/colours}

\item{mp_pal}{continuous colour palette for move persistence values}

\item{id_pal}{discrete colour palette for track id's}

\item{date_pal}{continuous colour palette for displaying date along track}
}
\description{
set aesthetics as a named list for: 1) mapping i) estimated
locations; ii) estimated confidence ellipses; iii) estimated track lines;
iv) observed locations; v) land masses; vi) water bodies; and 2) colour
palettes for i) behavioural indices or ii) individual animal tracks
}
