% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mpm.R
\name{fit_mpm}
\alias{fit_mpm}
\title{fit a \strong{Move Persistence Model} (\code{mpm})}
\usage{
fit_mpm(
  x,
  what = "predicted",
  model = c("jmpm", "mpm"),
  coords = 3:4,
  control = mpm_control(),
  inner.control = NULL,
  optim = NULL,
  optMeth = NULL,
  verbose = NULL
)
}
\arguments{
\item{x}{a \code{ssm_df} fit object or a data frame of observations (see details)}

\item{what}{if a \code{ssm_df} fit object is supplied then \code{what} determines
whether fitted, predicted (default), or rerouted values are mapped; ignored if
\code{x} is a data frame}

\item{model}{mpm model to fit; either \code{mpm} with unpooled random walk
variance parameters (\code{sigma_(g,i)}) or \code{jmpm} with a single,
pooled random variance parameter (\code{sigma_g})}

\item{coords}{column numbers of the location coordinates (default = 3:4)}

\item{control}{list of control settings for the outer optimizer
(see \link{mpm_control} for details)}

\item{inner.control}{list of control parameters for the inner optimization}

\item{optim}{is deprecated, use ssm_control(optim = "optim") instead,
see \link{ssm_control} for details}

\item{optMeth}{is deprecated, use ssm_control(method = "L-BFGS-B") instead,
see \link{ssm_control} for details}

\item{verbose}{is deprecated, use ssm_control(verbose = 1) instead,
see \link{ssm_control} for details}
}
\value{
a list with components
\itemize{
\item \code{fitted} a dataframe of fitted locations
\item \code{par} model parameter summary
\item \code{data} input data.frame
\item \code{tmb} the \code{TMB} object
\item \code{opt} the object returned by the optimizer
}
}
\description{
fit a random walk with time-varying move persistence to
temporally regular or irregular location data
}
\examples{
## fit jmpm to two southern elephant seal tracks
xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
control = ssm_control(verbose = 0))

fmpm <- fit_mpm(xs, model = "jmpm")

}
\references{
Jonsen ID, McMahon CR, Patterson TA, et al. (2019) Movement responses to
environment: fast inference of variation among southern elephant seals with
a mixed effects model. Ecology. 100(1):e02566
\url{https://doi.org/10.1002/ecy.2566}
}
