% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ssm.R
\name{fit_ssm}
\alias{fit_ssm}
\title{Fit Continuous-Time State-Space Models to filter Argos satellite
geolocation data}
\usage{
fit_ssm(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("lonerr", "laterr"),
  vmax = 5,
  ang = c(15, 25),
  distlim = c(2500, 5000),
  spdf = TRUE,
  min.dt = 60,
  pf = FALSE,
  model = "crw",
  time.step = NA,
  emf = NULL,
  map = NULL,
  parameters = NULL,
  fit.to.subset = TRUE,
  control = ssm_control(),
  inner.control = NULL,
  verbose = NULL,
  optim = NULL,
  optMeth = NULL,
  lpsi = NULL
)
}
\arguments{
\item{x}{a data frame of observations including Argos KF error ellipse info
(when present)}

\item{id}{the name (as a quoted character string) of id variable: a unique
identifier for individual (animal) track data sets.}

\item{date}{the name (as a quoted character string)of the date/time variable:
date and time (as YYYY-MM-DD HH:MM:SS) of each observation.}

\item{lc}{the name (as a quoted character string) of the location quality class
variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z").
Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS)
and other data types.}

\item{coord}{the names (as quoted character strings) of the location coordinate
variables: defaults are c("lon","lat"), but could also be c("x","y") for planar
coordinates; or if input data is an \code{sf} object then "geometry". If input
data is an \code{sf} object then \code{coord} is set to "geometry" by default.}

\item{epar}{the names (as quoted character strings) of the Argos error ellipse
parameters: defaults are "smaj" (ellipse semi-major axis),
"smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if
these variables are missing from the input data.}

\item{sderr}{the names (as quoted character strings) of provided standard
errors in longitude and latitude: defaults are "lonerr", "laterr". Typically,
these are only provided for processed light-level geolocation data. Ignored if
these variables are missing from the input data.}

\item{vmax}{max travel rate (m/s) passed to \link[trip:sda]{trip::sda} to identify
outlier locations}

\item{ang}{angles (deg) of outlier location "spikes"}

\item{distlim}{lengths (m) of outlier location "spikes"}

\item{spdf}{(logical) turn \link[trip:sda]{trip::sda} pre-filtering on (default; TRUE) or off}

\item{min.dt}{minimum allowable time difference between observations;
\code{dt <= min.dt} will be ignored by the SSM}

\item{pf}{just pre-filter the data, do not fit the SSM (default is FALSE)}

\item{model}{fit a simple random walk (\code{rw}), correlated random walk
(\code{crw}), or a time-varying move persistence model (\code{mp}), all as
continuous-time process models}

\item{time.step}{options: 1) the regular time interval, in hours, to predict to;
2) a vector of prediction times, possibly not regular, must be
specified as a data.frame with id and POSIXt dates; 3) NA - turns off
prediction and locations are only estimated at observation times.}

\item{emf}{optionally supplied data.frame of error multiplication factors for
Argos location quality classes. Default behaviour is to use the factors
supplied by \link{emf}}

\item{map}{a named list of parameters as factors that are to be fixed during
estimation, e.g., \code{list(psi = factor(NA))}}

\item{parameters}{a list of initial values for all model parameters and
unobserved states, default is to let sfilter specify these. Only play with
this if you know what you are doing}

\item{fit.to.subset}{fit the SSM to the data subset determined by
\link{prefilter} (default is TRUE)}

\item{control}{list of control settings for the outer optimizer
(see \link{ssm_control} for details)}

\item{inner.control}{list of control settings for the inner optimizer
(see \link[TMB:MakeADFun]{TMB::MakeADFun} for additional details)}

\item{verbose}{is deprecated, use ssm_control(verbose = 1) instead,
see \link{ssm_control} for details}

\item{optim}{is deprecated, use \code{ssm_control(optim = "optim")} instead,
see \link{ssm_control} for details}

\item{optMeth}{is deprecated, use \code{ssm_control(method = "L-BFGS-B")} instead,
see \link{ssm_control} for details}

\item{lpsi}{is deprecated, use \code{ssm_control(lower = list(lpsi = -Inf))} instead,
see \link{ssm_control} for details}
}
\value{
a list with components
\itemize{
\item \code{call} the matched call
\item \code{predicted} an sf tbl of predicted location states
\item \code{fitted} an sf tbl of fitted locations
\item \code{par} model parameter summary
\item \code{data} an augmented sf tbl of the input data
\item \code{inits} a list of initial values
\item \code{pm} the process model fit, either "rw" or "crw"
\item \code{ts} time time.step in h used
\item \code{opt} the object returned by the optimizer
\item \code{tmb} the TMB object
\item \code{rep} TMB sdreport
\item \code{aic} the calculated Akaike Information Criterion
\item \code{time} the processing time for sfilter
}
}
\description{
fits: i) a simple random walk (\code{rw}) ii) a correlated random walk
(\code{crw} - a random walk on velocity), or iii) a time-varying move persistence
model (\code{mp}), all in continuous-time, to filter Argos LS, and/or KF/KS
location data, processed light-level geolocation data (GLS), and/or GPS data.
Location data of different types can combined in a single data frame
(see details). Predicts locations at user-specified time intervals
(regular or irregular).
}
\details{
\code{x} is a \code{data.frame}, \code{tibble}, or \code{sf-tibble} with 5, 7 or 8
columns, depending on the tracking data type. Argos Least-Squares and GPS
data should have 5 columns in the following order:
\strong{\code{id}, \code{date}, \code{lc}, \code{lon}, \code{lat}}. Where \code{date} can be a POSIX object or text
string in YYYY-MM-DD HH:MM:SS format. If a text string is supplied then the
time zone is assumed to be \code{UTC}. lc (location class) can include the
following values: 3, 2, 1, 0, A, B, Z, G, or GL. The latter two are for GPS
and GLS locations, respectively. Class Z values are assumed to have the same
error variances as class B. By default, class \code{G} (GPS) locations are assumed
to have error variances 10x smaller than Argos class 3 variances, but unlike
Argos error variances the GPS variances are the same for longitude and latitude.

See \link{emf} for details on how to modify these assumptions.

Argos Kalman Filter (or Kalman Smoother) data should have 8 columns,
including the above 5 plus \strong{\code{smaj}, \code{smin}, \code{eor}} that contain Argos error
ellipse variables (in m for \code{smaj}, \code{smin} and deg for \code{eor}).

Light-level geolocation (GLS) locations can be modelled provided each
longitude and latitude has a corresponding standard error. These data should
have 7 columns, including the above 5 plus \code{lonerr}, \code{laterr} (in degrees).
In this case, all lc values should be set to \code{GL}.

Multiple location data types can be combined in a single data frame
(see the vignette for examples).

When data are provided as an \code{sf-tibble}, the user-specified projection is
respected. Otherwise, longlat data are re-projected internally to a global
Mercator grid and provided as the default output. A simple \code{tibble}, without
a geom, of \verb{lon,lat} and \verb{x,y} location estimates can be obtained by using
\link{grab} with the argument \code{as_sf = FALSE}.
}
\examples{
## fit crw model to Argos LS data
fit <- fit_ssm(ellie, vmax = 4, model = "crw", time.step = 24, control = ssm_control(verbose = 0)) 

## time series plots of fitted values and observations
plot(fit, what = "fitted", type = 1, ask = FALSE)

## 2-D tracks plots of predicted values and observations
plot(fit, what = "predicted", type = 2, ask = FALSE)


}
\references{
Jonsen ID, Patterson TA, Costa DP, et al. (2020) A continuous-time state-space
model for rapid quality-control of Argos locations from animal-borne tags.
Movement Ecology 8:31

Jonsen ID, McMahon CR, Patterson TA, et al. (2019) Movement responses to
environment: fast inference of variation among southern elephant seals with
a mixed effects model. Ecology. 100(1):e02566
}
