% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap.R
\name{fmap}
\alias{fmap}
\title{fmap}
\usage{
fmap(
  x,
  y = NULL,
  what = c("fitted", "predicted", "rerouted"),
  conf = TRUE,
  obs = FALSE,
  obs.shp = 17,
  by.id = TRUE,
  by.date = TRUE,
  crs = NULL,
  ext.rng = c(0.05, 0.05),
  size = 0.25,
  col.ssm = "dodgerblue",
  col.obs = "black",
  lines = FALSE,
  landfill = grey(0.6),
  map_type = "default",
  pal = "Cividis",
  rev = FALSE,
  last_loc = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{foieGras} ssm fit object with class \code{ssm_df}}

\item{y}{optionally, a \code{foieGras} mpm fit object with class \code{mpm_df};
default is NULL}

\item{what}{specify which location estimates to map: fitted, predicted or
rerouted}

\item{conf}{include confidence regions around estimated location (logical;
default = TRUE, unless y is an mpm fit object then conf is FALSE)}

\item{obs}{include Argos observations on map (logical; default = FALSE)}

\item{obs.shp}{point shape for observations (default = 17)}

\item{by.id}{when mapping multiple tracks, should locations be coloured by
id (logical; default = TRUE if nrow(x) > 1 else FALSE)}

\item{by.date}{when mapping single tracks, should locations be coloured by
date (logical; default = TRUE if nrow(x) == 1 else FALSE)}

\item{crs}{\code{proj4string} for re-projecting locations, if NULL the
default projection ("+proj=merc") for the fitting the SSM will be used}

\item{ext.rng}{factors to extend the plot range in x and y dimensions
(can exceed 1)}

\item{size}{size of estimated location points (size = NA will draw no points).
Optionally, a vector of length 2 with size of observed locations given by 2nd
value (ignored if obs = FALSE)}

\item{col.ssm}{colour of ssm-fitted or ssm-predicted locations (ignored if by.id = TRUE)}

\item{col.obs}{colour of observed locations (ignored if obs = FALSE)}

\item{lines}{logical indicating if lines are added to connect estimated
locations (default = FALSE)}

\item{landfill}{colour to use for land (default = grey(0.6))}

\item{map_type}{background map type (default = NULL, which uses rnaturalearth
to add landmasses); if packages \code{ggspatial} & \code{rosm} are installed
then any map type returned by \link[rosm:osm.types]{rosm::osm.types} can be used for a more
detailed map background.}

\item{pal}{\link{hcl.colors} palette to use (default: "Cividis"; type
\code{\link[=hcl.pals]{hcl.pals()}} for options)}

\item{rev}{reverse colour palette (logical)}

\item{last_loc}{colour to render last location of each track (default = NULL)}

\item{...}{additional arguments passed to \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile}}
}
\description{
map foieGras fitted or predicted locations, with or without
Argos observations, optionally apply a different projection
}
