% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap.R
\name{fmap}
\alias{fmap}
\title{fmap}
\usage{
fmap(x, what = c("fitted", "predicted"), obs = FALSE, crs = NULL,
  ext.rng = c(0.05, 0.05), size = 1)
}
\arguments{
\item{x}{a foieGras fitted object}

\item{what}{specify which location estimates to map: fitted or predicted}

\item{obs}{include Argos observations on map (logical)
(logical); ignored if `obs = FALSE`}

\item{crs}{`proj4string` for reprojecting locations, if NULL the
default projection ("+proj=merc") for the fitting the SSM will be used}

\item{ext.rng}{factors to extend the plot range in x and y dimensions
(can exceed 1)}

\item{size}{size of estimated location points}
}
\description{
map foieGras fitted or predicted locations, with or without
Argos observations, optionally apply a different projection
}
