% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-deprecated.R
\name{pluck}
\alias{pluck}
\alias{extract}
\title{Deprecated functions.}
\usage{
pluck(...)

extract(...)
}
\arguments{
\item{...}{ignored}
}
\description{
\code{extract} and \code{pluck} have been deprecated to avoid conflict with raster::extract, tidyr::extract, purrr::pluck. Instead use
\code{grab}.
}
