% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foieGras-package.R
\docType{package}
\name{foieGras-package}
\alias{foieGras-package}
\alias{foieGras}
\title{\pkg{foieGras}}
\description{
fit Continuous-Time Random Walk and Correlated Random Walk state-space models to filter Argos Least Squares or Kalman Filter location data
}
\references{
Jonsen I, McMahon CR, Patterson TA, Auger-Méthé M, Harcourt R, Hindell MA, Bestley S (2019) 
Movement responses to environment: fast inference of variation among southern elephant seals with a mixed effects model. 
Ecology 100:e02566

Jonsen I, Patterson TA, Costa DP, Doherty PD, Godley BJ, Grecian WJ, Guinet C, Hoenner X, Kienle SS, 
Robinson PW, Votier SC, Witt MJ, Hindell MA, Harcourt RG, McMahon CR (2020) A continuous-time state-space model for rapid 
quality-control of Argos locations from animal-borne tags. arXiv doi:arxiv:2005.00401
}
\seealso{
fit_ssm
}
\author{
Ian Jonsen, Toby Patterson
}
\keyword{foieGras}
