% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Coerce input data into expected foieGras format}
\usage{
format_data(
  x,
  id = "id",
  date = "date",
  lc = "lc",
  coord = c("lon", "lat"),
  epar = c("smaj", "smin", "eor"),
  sderr = c("lonerr", "laterr")
)
}
\arguments{
\item{x}{input data}

\item{id}{the name (as a quoted character string) of id variable: a unique
identifier for individual (animal) track data sets.}

\item{date}{the name (as a quoted character string)of the date/time variable:
date and time (as YYYY-MM-DD HH:MM:SS) of each observation.}

\item{lc}{the name (as a quoted character string) of the location quality class
variable: Argos location quality class (values in the set: 3,2,1,0,"A","B","Z").
Can also include "G" for GPS data and/or "GL" for light-level geolocation (GLS)
and other data types.}

\item{coord}{the names (as quoted character strings) of the location coordinate
variables: defaults are c("lon","lat"), but could also be c("x","y") for planar
coordinates; or if input data is an \code{sf} object then "geometry". If input
data is an \code{sf} object then \code{coord} is set to "geometry" by default.}

\item{epar}{the names (as quoted character strings) of the Argos error ellipse
parameters: defaults are "smaj" (ellipse semi-major axis),
"smin" (ellipse semi-minor axis), and "eor" (ellipse orientation). Ignored if
these variables are missing from the input data.}

\item{sderr}{the names (as quoted character strings) of provided standard
errors in longitude and latitude: defaults are "lonerr", "laterr". Typically,
these are only provided for processed light-level geolocation data. Ignored if
these variables are missing from the input data.}
}
\value{
a data.frame or sf-tibble of input data in expected foieGras format
}
\description{
\code{format_data}
}
\details{
called by \code{fit_ssm}.
}
\keyword{internal}
