% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck.R
\name{grab}
\alias{grab}
\title{grab tbl's by name from a foieGras fitted model object}
\usage{
grab(fitobj, what, as_sf = TRUE)
}
\arguments{
\item{what}{the tibble to be grabbed; either `fitted`, `predicted`, or `data` (single letters can be used)}

\item{as_sf}{logical; if FALSE then return a tibble with unprojected lonlat coordinates, otherwise return an sf tibble}

\item{...}{additional arguments are ignored}

\item{fitobj}{a foieGras fitted model object}
}
\value{
a tbl with all individual tbl's appended
}
\description{
`grab()` lets you obtain `fitted`, `predicted`, or `data` tbl's from
a compound tbl created when fitting to multiple individual data sets. The specified tbl's
are appended to a single output tbl.
}
\examples{
\dontrun{
data("ellie")
ellie.fit <- fit_ssm(ellie, model = "crw", time.step = 6)

pred <- pluck(ellie.fit, "predicted")
}
}
