% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\title{join an mpm-estimated behavioural index to ssm-predicted locations}
\usage{
join(ssm, mpm, as_sf = TRUE)
}
\arguments{
\item{ssm}{a \code{foieGras} ssm fitted model object}

\item{mpm}{a \code{foieGras} mpm fitted model object}

\item{as_sf}{logical; if FALSE then return a tibble with unprojected lonlat
coordinates, otherwise return an sf tibble}
}
\value{
a single tbl with all individuals
}
\description{
`join()` joins ssm-predicted locations and mpm-estimated behavioural index into a single tibble. If the ssm-predicted tibble is a projected sf object then the output of join will also be an sf object (default). This can be avoided by using `as_sf = FALSE`.
}
\examples{
## load example foieGras fit objects (to save time)
data(fssm)
data(fmpm)
## join predicted values as an unprojected tibble
fsmp <- join(fssm, fmpm, as_sf = FALSE)
fsmp
}
