% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(
  x,
  y = NULL,
  what = c("fitted", "predicted", "rerouted"),
  by.id = TRUE,
  by.date = TRUE,
  crs = NULL,
  ext.rng = c(0.05, 0.05),
  map_type = "default",
  aes = aes_lst(),
  normalise = FALSE,
  group = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{foieGras} ssm fit object with class \code{ssm_df}}

\item{y}{optionally, a \code{foieGras} mpm fit object with class \code{mpm_df}}

\item{what}{specify which location estimates to map: fitted, predicted or
rerouted}

\item{by.id}{when mapping multiple tracks, should locations be coloured by
id (logical; default = TRUE if nrow(x) > 1 else FALSE; ignored if behavioural
index provided)}

\item{by.date}{when mapping single tracks, should locations be coloured by
date (logical; default = TRUE if nrow(x) == 1 else FALSE; ignored if behavioural
index provided)}

\item{crs}{\code{proj4string} for re-projecting locations, if NULL the
default projection (Mercator) for the fitting the SSM will be used}

\item{ext.rng}{proportion to extend the plot range in x and y dimensions}

\item{map_type}{background map type (default = NULL, which uses rnaturalearth
to add landmasses); if packages \code{ggspatial} & \code{rosm} are installed
then any map type returned by \link[rosm:osm.types]{rosm::osm.types} can be used for a more
detailed coastline, given appropriate zoom settings (see
\link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile} for details).}

\item{aes}{a list of map aesthetics (shape, size, col, fill, alpha) for each
map feature (estimated locations, confidence ellipses, track lines, observed
locations, land masses, water bodies). Constructed by \code{aes_lst()} to be
applied, in order, to: 1) estimated locations; 2) confidence ellipses;
3) track lines; 4) observed locations; 5) land regions; 6) water regions}

\item{normalise}{logical; if output includes a move persistence estimate,
should g be normalised to 0, 1 (default = FALSE)}

\item{group}{logical; should move persistence be normalised among individuals
as a group (ie. relative) or separately (default = FALSE)}

\item{...}{additional arguments passed to \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile}}
}
\description{
map foieGras-estimated locations and behavioural indices with
coastline and projection options
}
