% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(
  x,
  y = NULL,
  what = c("fitted", "predicted", "rerouted"),
  aes = aes_lst(),
  by.id = TRUE,
  by.date = FALSE,
  crs = NULL,
  ext.rng = c(0.05, 0.05),
  map_type = "default",
  normalise = FALSE,
  group = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{foieGras} ssm fit object with class \code{ssm_df}}

\item{y}{optionally, a \code{foieGras} mpm fit object with class \code{mpm_df}}

\item{what}{specify which location estimates to map: fitted, predicted or
rerouted}

\item{aes}{a list of map controls and aesthetics (shape, size, col, fill, alpha)
for each map feature (estimated locations, confidence ellipses, track lines,
observed locations, land masses, water bodies). Constructed by \code{aes_lst()} and
can be modified for custom maps (see examples)}

\item{by.id}{when mapping multiple tracks, should locations be coloured by
id (logical; default = TRUE if \code{nrow(x) > 1} else FALSE; ignored if behavioural
index provided)}

\item{by.date}{when mapping single tracks, should locations be coloured by
date (logical; default = FALSE; ignored if behavioural
index provided)}

\item{crs}{\code{proj4string} for re-projecting locations, if NULL the
default projection (Mercator) for the fitting the SSM will be used}

\item{ext.rng}{proportion (can exceed 1) to extend the plot range in x and y
dimensions}

\item{map_type}{background map type ("default" uses \link{rnaturalearth}
to add landmasses). If \link{rnaturalearthdata} is installed; if packages \link{ggspatial}
and \link{rosm} are installed then any tile map type returned by \link[rosm:osm.types]{rosm::osm.types}
can be used for a potentially more detailed coastline at fine spatial scales,
given appropriate zoom settings (see \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile} for details).}

\item{normalise}{logical; if output includes a move persistence estimate,
should g (the move persistence index) be normalised to have minimum = 0 and
maximum = 1 (default = FALSE).}

\item{group}{logical; should g be normalised among individuals as a group,
a 'relative g', or separately to highlight regions of lowest and highest move
persistence along a track (default = FALSE).}

\item{...}{additional arguments passed to \link[ggspatial:annotation_map_tile]{ggspatial::annotation_map_tile}}
}
\value{
a map as a ggplot2 object
}
\description{
map foieGras-estimated locations and behavioural indices with
coastline and projection options
}
\examples{
## an ssm fit object
fit <- fit_ssm(sese1, model = "rw", time.step = 24, control = ssm_control(verbose = 0))

## redner default map
map(fit, what = "p")

## construct aes list to modify map
aes <- aes_lst()

## modify aes to turn on estimated track line & observed Argos locations 
aes$obs <- aes$line <- TRUE
## set ocean colour to pale blue
aes$df$fill[6] <- hcl.colors(n=1, palette = "Blues", alpha = 0.3)

## map using modified aes list, with polar stereographic projection centered
##    on approximate track midpoint, extend x,y limits by 10\%
m <- map(fit, what = "p", aes = aes, crs = "+proj=stere +lon_0=90 +units=km +datum=WGS84", ext.rng = c(0.1,0.1))

## use thinner graticule lines than ggplot2 default
m + theme(panel.grid = element_line(size = 0.1))


}
