% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osar.r
\name{osar}
\alias{osar}
\title{calculate one-step-ahead (prediction) residuals from a \code{aniMotum}
\code{ssm} fit}
\usage{
osar(x, method = "fullGaussian", ...)
}
\arguments{
\item{x}{a \code{aniMotum} \code{ssm} fit object with class \code{ssm_df}}

\item{method}{method to calculate prediction residuals
(default is \code{oneStepGaussianOffMode}; see \link[TMB:oneStepPredict]{TMB::oneStepPredict} for details)}

\item{...}{other arguments to \link[TMB:oneStepPredict]{TMB::oneStepPredict}}
}
\description{
calculate one-step-ahead (prediction) residuals from a \code{aniMotum}
\code{ssm} fit
}
\details{
One-step-ahead residuals are useful for assessing goodness-of-fit
in latent variable models. This is a wrapper function for \link[TMB:oneStepPredict]{TMB::oneStepPredict}
(beta version). \code{osar} tries the \code{fullGaussian} (fastest) method first and
falls back to the \code{oneStepGaussianOffMode} (slower) method for any failures.
Subsequent failures are dropped from the output and a warning message is given.
Note, OSA residuals can take a considerable time to calculate if there are
many individual fits and/or deployments are long. The method is automatically
parallelised across 2 x the number of individual fits, up to the number of
processor cores available.
}
\examples{
# generate a ssm fit object (call is for speed only)
xs <- fit_ssm(ellie, spdf=FALSE, model = "rw", time.step=24, control = ssm_control(verbose = 0))

res <- osar(xs)

}
\references{
Thygesen, U. H., C. M. Albertsen, C. W. Berg, K. Kristensen, and
A. Neilsen. 2017. Validation of ecological state space models using the Laplace
approximation. Environmental and Ecological Statistics 24:317–339.
}
