% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_sf_project.R
\name{pf_sf_project}
\alias{pf_sf_project}
\title{handle spatial projection}
\usage{
pf_sf_project(x)
}
\arguments{
\item{x}{data from \code{pf_sda_filter}}
}
\description{
project from longlat to merc or respect user-supplied
projection (if not longlat) & ensure that longitudes straddling -180,180 or
0,360 are shifted appropriately.
}
\keyword{internal}
