% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.foieGras.R
\name{plot.fG}
\alias{plot.fG}
\title{plot}
\usage{
\method{plot}{fG}(x, what = c("fitted", "predicted"), type = 1,
  ncol = 2, lc = FALSE, outlier = FALSE, ...)
}
\arguments{
\item{x}{a foieGras ssm compound tibble}

\item{what}{specify which location estimates to display on time-series plots: fitted or predicted}

\item{type}{of plot to generate: 1-d time series for lon and lat separately (type = 1, default) or 2-d track plot (type = 2)}

\item{ncol}{number of columns to use for facetting. Default is ncol = 2 but this may be increased for large compound fit objects}

\item{lc}{display Argos location classes}

\item{outlier}{include all extreme outliers flagged by prefilter in plots (logical)}

\item{...}{additional arguments to be ignored}
}
\description{
visualise multiple fits from an fG compound tibble
}
\examples{
## load example foieGras fit object (to save time)
data(fit)
plot(fit, what = "f", type = 1)
plot(fit, what = "p", type = 2, lc = TRUE)

}
