% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_osar.R
\name{plot.fG_osar}
\alias{plot.fG_osar}
\title{plot}
\usage{
\method{plot}{fG_osar}(x, type = c("qqnorm", "histogram"), bw = 0.5, ...)
}
\arguments{
\item{x}{a \code{foieGras osar} object with class `fG_osar`}

\item{type}{type of residual plot to generate; either qqnorm (default) or histogram}

\item{bw}{binwidth for histogram plots (see ggplot2::geom_histogram for details), ignored if type = "qqnorm"}

\item{...}{additional arguments to be ignored}
}
\description{
plot One-Step-Ahead (prediction) residuals from a \code{foieGras osar} object
}
\examples{
## load example osar output (to save time)
data(fssm)
d <- fssm[1, ] ## just use first seal to save time
dres <- osar(d)
plot(dres, type = "qq")

}
