% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fG_simfit.R
\name{plot.fG_simfit}
\alias{plot.fG_simfit}
\title{plot}
\usage{
\method{plot}{fG_simfit}(x, ncol = 1, pal = "Zissou1", ...)
}
\arguments{
\item{x}{a \code{foieGras} simulation data.frame with class \code{fG_simfit}}

\item{ncol}{number of columns to arrange multiple plots}

\item{pal}{\code{hcl.colors} palette to use (default: "Zissou1"; type \code{hcl.pals()} for options)}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of simulated tracks.
}
\description{
visualize tracks simulated from a foieGras model fit
}
\examples{
trs <- simfit(xs, what = "p", reps = 1)
plot(trs, ncol = 2)

}
