% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.foieGras.R
\name{plot.foieGras}
\alias{plot.foieGras}
\title{plot}
\usage{
\method{plot}{foieGras}(x, what = c("fitted", "predicted"),
  outlier = FALSE, ...)
}
\arguments{
\item{x}{a foieGras fitted object}

\item{what}{specify which location estimates to display on time-series plots: fitted or predicted}

\item{outlier}{include all extreme outliers flagged by prefilter in plots (logical)}

\item{...}{additional arguments to be ignored}
}
\description{
visualise foieGras SSM fits to track data
}
\examples{
## load example foieGras fit object (to save time)
data(fit)
plot(fit$ssm[[1]])

}
