% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.R
\name{plot.sim}
\alias{plot.sim}
\title{plot}
\usage{
\method{plot}{sim}(x, type = 2, error = FALSE, pal = "Cividis", rev = FALSE, ...)
}
\arguments{
\item{x}{a \code{foieGras} simulation data.frame with class \code{sim}}

\item{type}{either 1, a 1-D time-series of speed (if model is \code{rw} or \code{crw}
specified behavioural states) or move persistence (g; if model is \code{mp}); or
2 (default), a 2-D track with location coloured by move persistence (g; if
model = "mp")}

\item{error}{logical, plot locations with error (TRUE) or without. Ignored
in 1-D time-series plots}

\item{pal}{\link[grDevices:palettes]{grDevices::hcl.colors} palette to use (default: "Cividis");
see \code{\link[grDevices:palettes]{grDevices::hcl.pals()}} for options}

\item{rev}{reverse direction of colour palette; logical (default = FALSE)}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of simulated tracks. Can be rendered all on a single page
(pages = 1) or on separate pages (pages = 0).
}
\description{
visualize simulated tracks from a \code{sim} data.frame
}
\examples{
tr <- sim(N=200, model = "mp")
plot(tr)

}
