% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim_post.R
\name{plot.sim_post}
\alias{plot.sim_post}
\title{plot}
\usage{
\method{plot}{sim_post}(
  x,
  type = c("lines", "points", "both"),
  zoom = TRUE,
  ncol = 1,
  hires = FALSE,
  ortho = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{aniMotum} simulation data.frame with class \code{sim_fit}}

\item{type}{plots tracks as "line", "points" or "both" (default).}

\item{zoom}{logical; should map extent be defined by track extent (TRUE; default) or
should global map be drawn (FALSE).}

\item{ncol}{number of columns to arrange multiple plots}

\item{hires}{logical; use high-resolution coastline data. Attempts to use
high-res coastline data via \link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries} with \code{scale = 10},
if the \code{rnaturalearthhires} data package is installed. This extends the
plot rendering time so is set to FALSE by default, in which case
\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries} with \code{scale = 50} data are used.}

\item{ortho}{logical; use an orthographic projection centered on the track
starting location(s) (TRUE; default). An orthographic projection may be
optimal for high latitude tracks and/or tracks that traverse long distances.
If FALSE then a global Mercator projection is used.}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of posterior simulated tracks.
}
\description{
visualize tracks simulated from a \code{aniMotum} model fit
}
\examples{
fit <- fit_ssm(ellie, model = "crw", time.step = 24)
psim <- sim_post(fit, what = "p", reps = 2)
plot(psim)

}
