% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simfit.R
\name{plot.simfit}
\alias{plot.simfit}
\title{plot}
\usage{
\method{plot}{simfit}(
  x,
  type = c("lines", "points", "both"),
  zoom = FALSE,
  or = NULL,
  ncol = 1,
  pal = "Viridis",
  ...
)
}
\arguments{
\item{x}{a \code{foieGras} simulation data.frame with class \code{simfit}}

\item{type}{plots tracks as "line", "points" or "both" (default).}

\item{zoom}{logical; should map extent be defined by track extent (TRUE) or
should global map be drawn (FALSE; default).}

\item{or}{orientation of projected map, default is to centre on
start of fitted track (ignored if \code{mapproj} package is not installed).}

\item{ncol}{number of columns to arrange multiple plots}

\item{pal}{\link[grDevices:palettes]{grDevices::hcl.colors} palette to use (default: "Viridis"; see
\code{\link[grDevices:palettes]{grDevices::hcl.pals()}} for options)}

\item{...}{additional arguments to be ignored}
}
\value{
Plots of simulated tracks.
}
\description{
visualize tracks simulated from a \code{foieGras} model fit
}
\examples{
fit <- fit_ssm(ellie, vmax = 4, model = "crw", time.step = 24)
trs <- simfit(fit, what = "p", reps = 2)
plot(trs, type = "b")

}
