% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefilter.R
\name{prefilter}
\alias{prefilter}
\title{Prepare Argos data for fitting a ct ssm}
\usage{
prefilter(d, vmax = 10, min.dt = 1)
}
\arguments{
\item{d}{input data - must have 5 (LS), or 8 (KF) columns (see details)}

\item{vmax}{max travel rate (m/s) passed to argosfilter::sdafilter to define outlier locations}

\item{min.dt}{minimum allowable time difference between observations; dt < min.dt will be ignored by the SSM}
}
\description{
\code{prefilter()} (1) determines Argos data type (LS or KF); (2) converts dates to POSIXt;
identifies observations with duplicate dates; (3) orders observations in time;
(4) removes duplicate observations; (5) removes observations occurring within 60 s of one another (keeps first);
(6) shifts longitudes that straddle -180,180 to 0,360 and vice-versa; (7) projects lonlat coords to mercator x,y
coords (in km); (8) adds location error multiplication factors based on Argos location class (for type LS);
and (9) uses a loess smooth to identify potential outlier locations (by distance only) to be ignored when fitting
the \code{ctrw} model
}
\details{
Internal function
}
