% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.foieGras.R
\name{print.foieGras}
\alias{print.foieGras}
\title{Print \code{ctrw}}
\usage{
\method{print}{foieGras}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{ctrw} fit object}

\item{digits}{number of digits to use in display}

\item{...}{unused. For compatibility with the generic method.}
}
\description{
Print \code{ctrw}
}
