% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickmap.r
\name{quickmap}
\alias{quickmap}
\title{quickmap}
\usage{
quickmap(x, what = c("fitted", "predicted"), crs = NULL,
  ext.rng = 0.2)
}
\arguments{
\item{x}{a foieGras fitted object}

\item{what}{specify which location estimates to map: fitted or predicted}
}
\description{
map foieGras fitted or predicted locations
}
