% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_path.R
\name{route_path}
\alias{route_path}
\title{Re-route a movement path around land using the \code{pathroutr} package}
\usage{
route_path(
  x,
  what = c("fitted", "predicted"),
  map_scale = 50,
  dist = 50000,
  append = TRUE,
  ...
)
}
\arguments{
\item{x}{either a \code{ssm} fit object or
a \code{simfit} object containing simulated paths}

\item{what}{if using a \code{ssm} object should the fitted (typically
irregular in time) or predicted (typically regular in time) locations be
re-routed.}

\item{map_scale}{scale of rnaturalearth map to use for land mass: one of 110,
50 (default), or 10. Note that map_scale = 10 is only available if you have
the \code{rnaturalearthhires} package installed see:
\url{https://github.com/ropensci/rnaturalearthhires}}

\item{dist}{buffer distance (m) to add around track locations. The convex
hull of these buffered locations defines the size of land polygon used to
aid re-routing of points on land. Larger buffers can result in longer
computation times. See London (2020) for further details. The default buffer
distance is a constant 50000 m.}

\item{append}{should re-routed locations be appended to the \code{ssm}
(ssm fit) object (default = TRUE), or returned as a tibble.}

\item{...}{additional arguments passed to pathroutr::prt_visgraph}
}
\description{
A wrapper function that uses the \code{pathroutr} package
\url{https://jmlondon.github.io/pathroutr/}
to re-route movement paths that cross a land barrier. The current
implementation will take either the output from a \code{fit_ssm} model or the
simulations generated by \code{simfit}.
}
\details{
\code{route_path} uses \link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries} at the medium (50)
scale, by default, to generate a land barrier. For efficient computation,
\code{route_path} clips the polygons to the buffered bounds (set by \code{dist} (in m)
of the movement track(s).

When the input is a \code{ssm} object \code{route_path} can append the
re-routed path locations to the \code{ssm} (ssm fit) object. This is useful
when move persistence is to be estimated from the re-routed locations via
\code{fit_mpm}, or tracks are to be visualised with \code{map}. \code{route_path} can also
return a standalone \code{tibble} of the re-routed path with the same number of
locations as either the original fitted or predicted locations.

When the re-routed path is appended to the \code{ssm} object, the path can be
extracted using the \code{grab} function, e.g. \code{grab(fit, what = "rerouted")}.

When the input is a \code{simfit} object then \code{route_path} returns the same
object but with the locations within each simulation re-routed.

We recommend that users working on complex rerouting problems and/or
requiring higher resolution land barrier data work with the \code{pathroutr}
package directly by first exctracting foieGras-estimated locations with
\code{grab}. Higher resolution land barrier data (polygon shapefiles) must be
obtained independently.
}
\examples{
# if 'pathroutr' is installed then ok to use route_path()
if(requireNamespace("pathroutr", quietly = TRUE)) {
  fit <- fit_ssm(ellie, vmax = 4, model = "crw", time.step = 24)
  fit <- route_path(fit, what = "predicted")
  grab(fit, what = "rerouted")
}

}
\references{
Josh M. London. (2020) pathroutr: An R Package for (Re-)Routing Paths Around
Barriers (Version v0.2.1) \url{https://zenodo.org/record/5522909#.YnPxEy_b1qs}
}
