% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_path.R
\name{route_path}
\alias{route_path}
\title{Re-route a movement path around land using \code{pathroutr}}
\usage{
route_path(x, what = c("fitted", "predicted"))
}
\arguments{
\item{x}{either a \code{fG_ssm} model fit object or
a \code{fG_simfit} object containing simulated paths}

\item{what}{if using a \code{fG_ssm} object re-route the fitted (typically irregular in time)
or predicted (typically regular in time) locations}
}
\description{
A wrapper function that uses the \code{pathroutr} package to re-route
movement paths that cross a land barrier. The current implementation will take either
the output from a \code{fit_ssm} model or the simulations generated by \code{simfit}.
}
\details{
If the input is a \code{fG_ssm} object then \code{route_path} returns a \code{tibble}
with the same dimensions as either the \code{fG_ssm} fitted or predicted locations.
If the input is a \code{fG_simfit} object then \code{route_path} returns the same nested object
but with the locations within each simulation re-routed.
}
\examples{
fit <- fit_ssm(sese, vmax = 4, model = "crw", time.step = 24)
rrt_preds <- route_path(fit, what = "predicted")

trs <- simfit(fit, what = "predicted", reps = 5)
rrt_sims <- route_path(trs)

}
\references{
Josh M. London. (2020) pathroutr: An R Package for (Re-)Routing Paths Around Barriers (Version v0.2.1)
\url{https://doi.org/10.5281/zenodo.4321827}
}
