% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtnorm.R
\name{rtnorm}
\alias{rtnorm}
\title{rtnorm}
\usage{
rtnorm(n, mean = 0, sd = 1, l = -Inf, u = Inf)
}
\arguments{
\item{n}{number of random values to generate}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}

\item{l}{lower limit of distribution}

\item{u}{upper limit of distribution}
}
\description{
rtnorm
}
\details{
simulate values from a truncated normal distribution
}
\examples{
x <- rtnorm(10, l = -5, u = 5)
range(x)
}
\keyword{internal}
