% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.R
\name{sfilter}
\alias{sfilter}
\title{fit a C-T SSM to individual track data}
\usage{
sfilter(x, model = c("rw", "crw"), time.step = 1,
  fit.to.subset = TRUE, parameters = NULL, optim = c("nlminb",
  "optim"), verbose = FALSE, inner.control = NULL)
}
\description{
Filter Argos LS or KF track data
}
\details{
Internal function
}
