% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfilter.R, R/sfilter_crw.r, R/sfilter_rw.R
\name{sfilter}
\alias{sfilter}
\title{fit a C-T SSM to individual track data}
\usage{
sfilter(d, ptime = 1, psi = 0, fit.to.subset = TRUE,
  parameters = NULL, optim = c("nlminb", "optim"), verbose = FALSE,
  inner.control = NULL)

sfilter(d, ptime = 1, psi = 0, fit.to.subset = TRUE,
  parameters = NULL, optim = c("nlminb", "optim"), verbose = FALSE,
  inner.control = NULL)

sfilter(d, ptime = 1, psi = 0, fit.to.subset = TRUE,
  parameters = NULL, optim = c("nlminb", "optim"), verbose = FALSE,
  inner.control = NULL)
}
\description{
Filter Argos LS or KF track data

Filter Argos LS or KF track data

Filter Argos LS or KF track data
}
\details{
Internal function

Internal function

Internal function
}
