% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_filter.R
\name{sim_filter}
\alias{sim_filter}
\title{Filter \code{simfit} simulations}
\usage{
sim_filter(trs, keep = 0.25, flag = 2)
}
\arguments{
\item{trs}{a \code{simfit} object}

\item{keep}{the quantile of flag values to retain}

\item{flag}{the similarity flag method (see details)}
}
\value{
a \code{simfit} object containing the filtered paths
}
\description{
This function calculates the similarity between the simulations
generated by \code{simfit} and the observed path used in the \code{ssm} fit,
and returns a \code{simfit} object containing the most similar tracks based on
a user specified quantile.
}
\details{
\itemize{
\item \code{flag = 1} will use an index based on Hazen (2017)
\item \code{flag = 2} (the default) will use a custom index
}
}
\examples{
## fit crw model to Argos LS data
fit <- fit_ssm(ellie, vmax = 4, model = "crw", time.step = 48)

## generate 10 simulated paths from ssm fit
trs <- simfit(fit, what = "predicted", reps = 10)

## filter simulations and keep paths in top 25\% of flag values
trs_f <- sim_filter(trs, keep = .33, flag = 2)

## compare unfiltered and filtered simulated paths
plot(trs) | plot(trs_f)

}
\references{
Hazen et al. (2017) WhaleWatch: a dynamic management tool for
predicting blue whale density in the California Current J. Appl. Ecol. 54: 1415-1428
\url{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/1365-2664.12820}
}
