% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_post.R
\name{sim_post}
\alias{sim_post}
\title{simulate from the posterior of a \code{ssm} fit, conditional on parameters and data}
\usage{
sim_post(x, reps = 1, what = "predicted")
}
\arguments{
\item{x}{a \code{ssm} fit object with class \code{ssm_df}}

\item{reps}{number of replicate tracks to simulate from an \code{ssm} model
fit object}

\item{what}{simulate fitted or predicted locations}
}
\value{
a \code{fG_sim_post} object containing the paths simulated from a
\code{ssm} fit object
}
\description{
...
}
\examples{
fit <- fit_ssm(sese2, model = "crw", time.step = 12)
psims <- sim_post(fit[2,], reps = 500, what = "pred")
plot(y ~ x, grab(fit[2,], "pred"), pch = NA)
sapply(1:length(psims), function(i) lines(psims[[i]], lwd=0.25, col = rgb(0.7,0.7,0.7,0.4)))
points(y ~ x, grab(fit[2,], "p"), pch = 19, cex=0.3, col="red")
points(y ~ x, grab(fit[2,], "d"), pch = 19, cex=0.3, col="orange")

}
