% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_flag.R
\name{similarity_flag}
\alias{similarity_flag}
\title{Add similarity flag value to a \code{sim_fit} object}
\usage{
similarity_flag(track, sim_track, flag = 2, cpf)
}
\arguments{
\item{track}{a dataframe containing longitude and latitudes of the observed
path used in the SSM fit}

\item{sim_track}{a dataframe containing the longitude and latitude of a single
simulated path from sim_fit}

\item{flag}{set to either 1 or 2 to define similarity index (see details)}

\item{cpf}{logical; are tracks simulated as central place foraging trips. If
cpf = TRUE then distance & bearing are calculated from the first location to
the most distant location, otherwise distance & bearing are calculated from
the first to the last location on each track.}
}
\value{
a single value representing the similarity in distance and bearing between the two paths
}
\description{
Calculate the similarity between a single path generated by sim_fit and the
observed path used in the SSM fit. In this context, similarity is calculated
as the sum of normalised differences in net displacement (km) and overall
bearing (deg) between the SSM-estimated path and the simulated paths. Similarity
flag values are based on two interpretations of the difference in overall distance
and bearing between the two paths.
}
\details{
\code{flag = 1} will use an index based on Hazen (2017)\cr
\code{flag = 2} (the default) will use a custom index
}
\keyword{internal}
