% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate animal tracks from a \code{fG_ssm} fit or from scratch}
\usage{
simulate(
  x = NULL,
  N = 100,
  start = list(c(0, 0), Sys.time()),
  model = c("rw", "crw", "mpm"),
  vmax = 4,
  sigma = c(4, 4),
  rho_p = 0,
  D = 0.05,
  tau = c(1.5, 0.75),
  rho_o = 0,
  sigma_g = 1.25,
  error = c("ls", "kf", "dist"),
  t_dist = c("reg", "gamma"),
  tpar = c(1.5, 0.5),
  alpha = c(0.9, 0.8),
  beta = c(1, 0.25)
)
}
\arguments{
\item{x}{a compound \code{fG} tbl fit object (ignored if NULL)}

\item{N}{number of time steps to simulate (ignored if x is suppled)}

\item{model}{simulate from the \code{rw} or \code{crw} process model (ignored if x is suppled)}

\item{error}{simulate Argos errors based on Least-Squares location classes ("lc"), Argos Kalman Filter error ellipses ("kf"), or from an arbitrary distribution ("dist") (ignored if x is suppled)}

\item{Sigma}{covariance matrix for movement process (ignored if x is suppled)}
}
\value{
a data.frame
}
\description{
simulate from the \code{rw} or \code{crw} process models to generate either a set of x,y (or lon,lat) coordinates from a \code{fG_ssm} fit with length equal to the number of observations used in the SSM fit, or a set of x,y (or lon,lat) coordinates with or without error from supplied input parameters.
}
\examples{

}
