% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{wrap_lon}
\alias{wrap_lon}
\title{Utility functions}
\usage{
wrap_lon(lon, lon_min = -180)
}
\arguments{
\item{lon}{a vector of longitudes}

\item{lon_min}{the minimum longitude value to wrap appropriately, eg. 0 to
wrap -180, 180 on to 0, 360 and -180 to wrap 0,360 on to -180,180}
}
\description{
Utility functions
}
\details{
called by prefilter
}
\examples{
lon <- seq(-180,180)
lon1 <- wrap_lon(lon, 0)
range(lon)
range(lon1)
}
